Update Lesson 08 materials: restructure and rename Tkinter example files for clarity
- Removed outdated `L08_example01_basic_window.py`.
- Added `L08_example01_basic_window_with_label.py` for better clarity in demonstration.
- Introduced `L08_example02_layout_managers.py` showcasing widget placement using Tkinter.
- Renamed and updated other example files for consistency: `L08_example02_buttons.py` -> `L08_example03_buttons.py`, `L08_example03_entry_fields.py` -> `L08_example04_entry_fields.py`, `L08_example04_temperature_converter.py` -> `L08_example05_temperature_converter.py`, `L08_example05_simple_game.py` -> `L08_example06_simple_game.py`, `L08_example06_hangman_gui.py` -> `L08_example07_hangman_gui.py`.
- Expanded nomenclature for key Tkinter concepts.
