+++ Lessons/lesson08/L08_example06_simple_game.py
