+++ Lessons/lesson08/L08_example07_hangman_gui.py
