+++ Lessons/lesson08/L08_nomenclature.txt
@@ -1,40 +1,84 @@
-LESSON 08 NOMENCLATURE - SIMPLE GRAPHICAL INTERFACES WITH TKINTER
+Nomenklatur för lektion 7:
 
-GUI (Graphical User Interface): 
-A visual way for users to interact with software using graphical elements instead of text commands.
+Det är väldigt många nya termer den här lektionen, återvänd gärna till denna fil
+	och använd den som en ordlista.
 
-Tkinter:
-Python's standard library for creating GUI applications.
+
+GUI (Graphical User Interface / Grafiskt Användargränssnitt):
+	Ett visuellt sätt för användare att interagera med program. Används istället
+		för textkommandon.
+
+tkinter:
+	En del av Pythons standardbibliotek (de saker som kommer färdiginstallerade
+		med Python). Används för att skapa GUI.
+	Namnet kommer ifrån att det är ett sätt att interagera med ett
+		verktygspaket som heter Tk. "Tk interface" helt enkelt.
 
 Widget:
-A GUI element like a button, label, or text field that users can interact with.
+	Ett GUI-element.
+		Exempel inkluderar knappar, textfält och ramar.
+	I tkinter så tillhör alltid en Widget en annan Widget. Detta skapar en
+		hierarki, som kan tänkas på som ett träd, där det finns Widgets i
+		Widgets i Widgets och så vidare.
+		Det enda undantaget är den Widget som man kallar för "root".
 
-Window:
-The main container that holds all other GUI elements.
+Window / Toplevel / Fönster:
+	Varje program som använder tkinter består av minst ett fönster. Fönstret
+		innehåller alla andra grafiska element.
+	Fönster är också Widgets.
 
-Label:
-A widget that displays text or images.
+Label / Etikett:
+	En Widget som kan visa text eller bilder.
 
-Button:
-A widget that users can click to trigger an action.
+Button / Knapp:
+	En Widget som användare kan trycka på för att få något att hända.
 
-Entry:
-A widget that allows users to input text.
+Entry / Textfält:
+	En Widget för att skriva in text.
 
-Frame:
-A container widget used to organize other widgets.
+Frame / Ram:
+	En Widget man använder som "behållare" för att organisera andra Widgets.
 
-Event:
-An action like a button click or key press that a program can respond to.
+Event / Händelse:
+	En reaktion från programmet på att användaren har gjort något eller
+		att programmet registrerar en förändring.
+	Ofta är dessa reaktioner på att användaren har klickat på något eller
+		tryckt på en tangent.
 
-Event Handler:
-A function that executes when a specific event occurs.
+Event Handler / Händelsehanterare:
+	En funktion som körs när ett särskilt Event sker.
 
-Layout Manager:
-A system for arranging widgets in a window (pack, grid, place).
+Layout Manager / Geometry Manager / Layouthanterare:
+	Ett system för att organisera och placera Widgets i en Widget som går att
+		använda som behållare. Exempelvis ett fönster eller en Frame.
+	I tkinter finns pack, grid och place.
 
-Mainloop:
-The method that starts a Tkinter application and waits for user interaction.
+mainloop:
+	Den metod som startar ett tkinter-baserat program och väntar på interaktion
+		från användaren.
+	Finns i det objekt som borde namngivits som root.
+	Utan denna så händer ingenting i tkinter om man startar programmet utanför
+		IDLE.
 
-Root:
-The main window of a Tkinter application.
+root / tk.Tk():
+	Huvudfönstret för en tkinter-applikation.
+		root är ett namn som är standard att använda, precis som man använder
+			argumentet "self" i metoder.
+	Stänger man root så stänger man alla Widgets som finns där i och avslutar
+		därmed automatiskt programmet.
+
+master:
+	Den Widget som en annan Widget tillhör. Det går att referera till en
+		Widgets master genom att använda syntaxen:
+			orginal_widgeten.master
+
+	En Widget kan bara ha en master.
+	En Widget kan vara master för flera Widgets.
+	Det är så man får den trädliknande hierarkin som nämndes tidigare.
+
+config:
+	Förkortning av configure (båda går att använda). Betyder konfigurera och
+		är en metod som finns på alla Widget.
+	Använder man den från en Widget så kan man ändra på egenskaper såsom text
+		och färg. Alla egenskaper utöver "name" går att ändra på. "name" kan
+		enbart sättas när en Widget skapas.
